+++ src/less-strict.ts
@@ -1,4 +1,6 @@
-export = {
+import type { Linter } from 'eslint';
+
+const config: Linter.Config = {
     rules: {
         'no-empty': 'off',
         'no-unused-vars': 'off',
@@ -6,14 +8,17 @@ export = {
         '@typescript-eslint/no-explicit-any': 'off',
         '@typescript-eslint/no-inferrable-types': 'off',
         '@typescript-eslint/no-empty-interface': 'off',
-        '@typescript-eslint/array-type': 'off',
         '@typescript-eslint/no-empty-function': 'off',
-        '@typescript-eslint/unbound-method': 'off',
-        '@typescript-eslint/require-await': 'off',
-        '@typescript-eslint/ban-ts-comment': 'off',
+        '@typescript-eslint/no-unsafe-argument': 'off',
         '@typescript-eslint/no-unsafe-call': 'off',
         '@typescript-eslint/no-unsafe-return': 'off',
         '@typescript-eslint/no-unsafe-member-access': 'off',
         '@typescript-eslint/no-unsafe-assignment': 'off',
+        '@typescript-eslint/array-type': 'off',
+        '@typescript-eslint/unbound-method': 'off',
+        '@typescript-eslint/require-await': 'off',
+        '@typescript-eslint/ban-ts-comment': 'off',
     },
 };
+
+export = config;