+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.2.0](https://github.com/goparrot/eslint-config/compare/v1.1.2...v1.2.0) (2024-06-12)
+
+
+### Styles
+
+* specify config typings ([fdd1dbf](https://github.com/goparrot/eslint-config/commit/fdd1dbf7fd3b18bc9f2cf6b754fdbd6c605654c1))
+
+### [1.0.3](https://github.com/goparrot/eslint-config/compare/v1.0.2...v1.0.3) (2021-12-16)
+
+### Bug Fixes
+
+*   turn off @typescript-eslint/no-unsafe-argument for less-strict ([8c6ee47](https://github.com/goparrot/eslint-config/commit/8c6ee47e1dfe2437bbdcaad62599414c550e99e3))
+
+### [1.0.2](https://github.com/goparrot/eslint-config/compare/v1.0.1...v1.0.2) (2021-12-10)
+
+### [1.0.2](https://github.com/goparrot/eslint-config/compare/v1.0.1...v1.0.2) (2021-12-07)
+
 ### [1.0.1](https://github.com/goparrot/eslint-config/compare/v1.0.0...v1.0.1) (2021-05-14)
 
 ### Bug Fixes